修正编译问题
