+++ code/client/rule/vnc/worker/attach_xx.go
@@ -1,0 +1,8 @@
+//go:build !windows
+// +build !windows
+
+package worker
+
+func attachDesktop() (func(), error) {
+	return func() {}, nil
+}
